+++ sdk/storage/azure_storage_blob/src/clients/blob_service_client.rs
@@ -16,7 +16,7 @@ use crate::{
 use azure_core::{
     credentials::TokenCredential,
     http::{
-        policies::{BearerTokenCredentialPolicy, Policy},
+        policies::{BearerTokenAuthorizationPolicy, Policy},
         NoFormat, PageIterator, RequestContent, Response, Url, XmlFormat,
     },
     Result,