+++ sdk/storage/azure_storage_blob/src/clients/page_blob_client.rs
@@ -18,7 +18,7 @@ use crate::{
 use azure_core::{
     credentials::TokenCredential,
     http::{
-        policies::{BearerTokenCredentialPolicy, Policy},
+        policies::{BearerTokenAuthorizationPolicy, Policy},
         NoFormat, RequestContent, Response, Url, XmlFormat,
     },
     Bytes, Result,